upgpkg: miniserve
There are multiple ring denpendencies in miniserve. And the `[patch]`
manifest section can help us overwrite all the dependency.

This patch declare that we're patching the source crates.io. This can
effectively add the git source of `ring` to the crates.io registry.
Whenever crates.io is queried for versions of `ring` it will return the
git source.

Reference:
----------
* https://doc.rust-lang.org/cargo/reference/overriding-dependencies.html

Signed-off-by: Avimitin <avimitin@gmail.com>