addpkg: av1an
This patch will fix two dependencies sources:

* Replace ffmpeg-next with my patched version

The `ffmpeg-next` crate relies on the `ffmpeg-sys-next` crate. And the
`ffmpeg-sys-next` misuses the arch type that LLVM can't recognize.

I've fixed this and made a PR for the `ffmpeg-sys-next`, but the
upstream is not active for nearly three months. So I have to replace
the crates.io source with my patched repository.

Reference: Avimitin/rust-ffmpeg-sys@3e2b72b

* Update serde to 1.0.132

The 1.0.132 version enables atomic64 serialization on RISC-V. And av1an
relies on this feature.

Reference: serde-rs/serde@0508cb5

Additional information: We must precisely update the `serde` and
`ffmpeg-next` crate. This may seem redundant but one of the dependencies
of the av1an, indicatif, didn't follow semver convention and make
breaking changes in the same major version. So we should not simply run
`cargo update` here.

Signed-off-by: Avimitin <avimitin@gmail.com>