addpkg: gdk-pixbuf2
This patch add gdk-pixbuf2 into noqemu list and increase the timeout
limit for "pixbuf-randomly-modified" test.

This package can't be built in qemu user environment. The
"pixbuf-randomly-modified" test randomly modifies the file and then
tries to load it. This test process needs vast amounts of RAM and then
the test will be killed by the SIGKILL signal.

It can be built successfully on Unmatched board, but we still can't pass
the "picbuf-randomly-modified" test. This time the failure is caused by
timeout. 500s is a reasonable value for the Unmatched performance.

Signed-off-by: Avimitin <avimitin@gmail.com>