addpkg: hck
This patch fix two error. One is the rustc could not find specification
for target "riscv64-unknown-linux-gnu" error.

Another one is related to the libz-sys version. Zlib fixes cmake
arch detection in [zlib-ng #942], and the crate libz-sys update the zlib
binding in [version 1.1.4]. However, the hck is using the 1.1.3 version
and causing itself can't be compiled in RISC-V.

Reference:
* [zlib-ng #942]: https://github.com/zlib-ng/zlib-ng/commit/81f1c8a41bd1427707f402450de8820816760fca
* [version 1.1.4]: https://github.com/rust-lang/libz-sys/commit/5c4704efa6ae73a95c3c6e1ccf4bd212f927be7a

Signed-off-by: Avimitin <avimitin@gmail.com>