addpkg: diskonaut
This patch fix three error:

1. The [target specification issue].

2. Crate libc can not be compiled due to symbol missing.

This is fixed in [rust-lang/libc PR #2668], and release in 0.2.118
version.

3. Test failure

Some tests of diskonaut depend on the previous events. It take a
snapshot of the TUI pixels output, then compare those snapshot pixel by
pixel to determine if the cli is working correctly.

However those test are not reliable. If the previous event took too
long to finish, the TUI output will not match what is expected.

The current workaround is to increase the event interval. One second
interval is enough for all event to finish its jobs.

Tested in 5950x QEMU environment.

Ref:
* [target specification issue]: https://github.com/felixonmars/archriscv-packages/issues/670#issuecomment-1025646326
* [rust-lang/libc PR #2668]: https://github.com/rust-lang/libc/pull/2668

Signed-off-by: Avimitin <avimitin@gmail.com>