addpkg: coreutils
This package fails to build on a riscv64 real hardware due to the same
reason on x86_64. The test-getlogin doesn't handle the case
/proc/self/loginuid is -1. After being reported to the upstream gnulib,
a patch to resolve this issue has been merged on master.
Ref: https://lists.gnu.org/archive/html/bug-gnulib/2022-06/msg00001.html

I also filed a bug on ArchLinux bug tracker which backports the patch to
resolve the issue of FTBFS.
Ref: https://bugs.archlinux.org/task/75112

Currently, I'm not sure why some unit tests fail when running with
qemu-user. Failing tests are listed below.
- FAIL: tests/misc/tac-2-nonseekable.sh
- FAIL: tests/tail-2/follow-stdin.sh