addpatch: xpra
Xpra consier all warning as Error. And ffmpeg has deprecate the usage of
the channel field in AvCodecContext struct. So this package will FTBFS.
Xpra upstream fix this issue in the master branch, but not tagged yet.

This PR upgrade the xpra to v4.3.4 (It's been flagged out-of-date on 2022-06-21),
and backport the ffmpeg warning fix.

Arch Linux upstream report: https://bugs.archlinux.org/task/76030?project=5&string=xpra

Signed-off-by: Avimitin <avimitin@gmail.com>