addpatch: bitwarden-cli
This patch remove all the electron dependencies and the husky installer.

Bitwarden-cli is part of the bitwarden/client project. It share
dependencies with other application. And the electron dependencies is
for desktop/Android version. Besides, electron is currently unusable on
RISC-V platform.

The husky installer is a git hook util implemented by nodejs. It is only
used for dev environment.

Signed-off-by: Avimitin <avimitin@gmail.com>