updpatch: cargo-msrv
The package cargo-msrv has hard-coded rustc version into its tests to
feed rustup, but unfortunately, this version does not have RISC-V
support, causing all tests that rely on rustup to fail.

This issue has been reported to the upstream, for now we should just
ignore those tests and wait for the upstream to develop a new MSRV
searching implementation.

Signed-off-by: Avimitin <avimitin@gmail.com>