devtools-riscv64: restore force architecture
`pacman` actually checks for valid architecture, and uses the output of
`uname()` for `auto`. This didn't break our regular packaging because
pacman only checks for the specified targets to install, and `base` /
`base-devel` are `any` packages and passes the check. It unfortunately
breaks when, for example, trying to run `pacstrap` with it as an
alternative pacman conf and specified anything that isn't an `any`
package.

While it may sound an out-of-scope usage, I think we should not rely on
the undocument fact that architecture checks are only done for the
specified targets. `handle->architectures` being `x86_64` only is a
problem on its own and _might_ break elsewhere too. So let's just be
specific here.