addpatch: virtiofsd
The virtiofsd package uses an unimplemented syscall `renameat` in riscv64
which causes build failure. This commit adds a compile condition to
exclude the target arch `riscv64` from building the binding to the
unsupported syscall. This commit also removes the `--target` option when
fetching source code to fix the target triple issue.

The additional compile option has been submitted to the upstream. The MR
status can be tracked at:
https://gitlab.com/virtio-fs/virtiofsd/-/merge_requests/156.

Signed-off-by: Avimitin <dev@avimit.in>