addpatch: mattermost
This package can be built on qemu-user without patch because in
`prepare()`, npm installs a static x86_64 pngquant binary that
can be executed in qemu-user.

But on real riscv64 board, this packae FTBFS because of missing
imagemin-pngquant dependency. Adding libpng to `makedepends`
resolves this issue because pngquant will be built from
source and dynamically linked.

libpng is only used when building the package so there is no need
to put it into `depends`.