Kernel updates with CONFIG_SERIAL_8250_DW=y
`8250_dw` needs to be initialized before the entry of an initrd to be
used as an initial console. This makes it necessary to be a builtin
instead of a module. Since it is common enough and present on multiple
boards, let's make this change into the common `riscv.defconfig`.

Also update existing kernels with this change:

- linux 6.6.8.arch1-1
- linux-zen 6.6.8.zen1-1
- linux-sophgo 6.1.61-2
- linux-th1520 6.6.8-1