addpatch: python-into-dbus
Test `test_unpacking` and `test_exceptions` are flaky:

```
Unreliable test timings! On an initial run, this test took 274.95ms, which exceeded the deadline of 200.00ms, but on a subsequent run it took 165.32 ms, which did not. If you expect this sort of variability in your test timings, consider turning deadlines off for this test by setting deadline=None.
```

Extend their deadline from the default 200ms to 1000ms.