addpatch: chezmoi 2.57.0-1
Disable the test timeout, it's passed by args instead of hard encoding in the source, so may not report upstream.

It costs ~800s on `centiskorch` to finish, on qemu-user it will costs more than 1 hour (1 hour is the timeout set on upstream action CI) and I'm not able to wait it to finish, each step costs several minutes. But I believe it finally could be before the universe death hh

 Just be patient, we are using RISC-V.