addpatch: libpeas-2 2.0.5-2
The test costs around 100s on both qemu-user and SG2042

Additionally, meson itself has a default timeout for test that is 30s, and the upstream source is not hard encoding any timeout value, just keep it as meson's default. The timeout could be passed by args while building, so we may not need to report it upstream.